Add files via upload
What’s Up?, an application that provides end to end message transmission with increased security using concepts of inbuilt encryption and decryption along with encoding and decoding. The application is developed using Python and QT for graphical user interface. Efficient algorithm and security improvisation constitute the focus of the application designed.